# artifacts/api-server/src/lib/botState.ts
export type BotPhase =
  | "idle"
  | "waiting_open"
  | "building_range"
  | "watching"
  | "in_trade"
  | "closed";

export type BreakoutDirection = "long" | "short" | "none";

export interface SessionState {
  symbol: string;
  phase: BotPhase;
  date: string;
  orbHigh: number | null;
  orbLow: number | null;
  orbWidth: number | null;
  currentPrice: number | null;
  volume: number | null;
  averageVolume: number | null;
  volumeRatio: number | null;
  openRangeStart: string | null;
  openRangeEnd: string | null;
  breakoutDirection: BreakoutDirection | null;
  entryPrice: number | null;
  stopPrice: number | null;
  targetPrice: number | null;
  currentPnl: number | null;
  longTradeUsed: boolean;
  shortTradeUsed: boolean;
  qty: number | null;
  alpacaOrderId: string | null;
}

// Per-bot config snapshot — a copy of global config that can evolve independently
export interface ChildBotConfig {
  openingRangeMinutes: number;
  riskPercent: number;
  rewardRiskRatio: number;
  requireVolumeConfirmation: boolean;
  volumeMultiplier: number;
  trailingStopEnabled: boolean;
  trailingStopActivationR: number;
  reEntryEnabled: boolean;
  maxOrbWidthPercent: number;
  minOrbWidthPercent: number;
  breakoutWindowMinutes: number;
  useModerateRisk: boolean;
}

// State for a single child bot instance
export interface ChildBotState {
  id: string;           // runtime id (e.g. "SPY-1")
  dbId: number | null;  // bot_instances row id
  symbol: string;
  generation: number;
  parentId: string | null;
  config: ChildBotConfig;
  session: SessionState;
  startedAt: string;
  stoppedAt: string | null;
  lastUpdated: string | null;
  error: string | null;
  // Performance tracking
  totalTrades: number;
  wins: number;
  losses: number;
  totalPnl: number;
  avgRMultiple: number;
  recentRMultiples: number[]; // last N r-multiples for evolution
  loopTimer: ReturnType<typeof setInterval> | null;
}

// The global registry of all running child bots
export interface BotRegistry {
  bots: Map<string, ChildBotState>;
  nextId: number;
}

export const registry: BotRegistry = {
  bots: new Map(),
  nextId: 1,
};

// Legacy single-bot state (kept for backward compat with /bot/* routes)
export interface BotState {
  running: boolean;
  symbol: string;
  phase: BotPhase;
  startedAt: string | null;
  stoppedAt: string | null;
  lastUpdated: string | null;
  error: string | null;
  session: SessionState | null;
  loopTimer: ReturnType<typeof setInterval> | null;
}

const today = (): string => new Date().toISOString().split("T")[0]!;

export const createEmptySession = (symbol: string): SessionState => ({
  symbol,
  phase: "waiting_open",
  date: today(),
  orbHigh: null,
  orbLow: null,
  orbWidth: null,
  currentPrice: null,
  volume: null,
  averageVolume: null,
  volumeRatio: null,
  openRangeStart: null,
  openRangeEnd: null,
  breakoutDirection: null,
  entryPrice: null,
  stopPrice: null,
  targetPrice: null,
  currentPnl: null,
  longTradeUsed: false,
  shortTradeUsed: false,
  qty: null,
  alpacaOrderId: null,
});

// Legacy single-bot state (wraps first bot in registry for backward compat)
export const botState: BotState = {
  running: false,
  symbol: "",
  phase: "idle",
  startedAt: null,
  stoppedAt: null,
  lastUpdated: null,
  error: null,
  session: null,
  loopTimer: null,
};