# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Opening Range Breakout (ORB) Trading Bot API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bots:
    get:
      operationId: listBots
      tags: [bots]
      summary: List all running child bots
      responses:
        "200":
          description: List of child bots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChildBot"
    post:
      operationId: createBot
      tags: [bots]
      summary: Start a new child bot for a symbol
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBotBody"
      responses:
        "200":
          description: Child bot started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChildBot"
    delete:
      operationId: stopAllBots
      tags: [bots]
      summary: Stop all running child bots
      responses:
        "200":
          description: All bots stopped
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /bots/{id}:
    get:
      operationId: getBot
      tags: [bots]
      summary: Get a single child bot by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Child bot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChildBot"
        "404":
          description: Not found
    delete:
      operationId: stopBot2
      tags: [bots]
      summary: Stop a single child bot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Bot stopped
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "404":
          description: Not found
  /bots/{id}/offspring:
    post:
      operationId: spawnOffspring
      tags: [bots]
      summary: Spawn an evolved offspring of a child bot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Offspring bot started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChildBot"
        "404":
          description: Parent bot not found
...
  /scanner/run:
    post:
      operationId: runScan
      tags: [scanner]
      summary: Trigger a manual gap scan
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScanRequest"
      responses:
        "200":
          description: Scan results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanResult"
...
components:
  schemas:
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
...
    CreateBotBody:
      type: object
      required: [symbol]
      properties:
        symbol:
          type: string
        config:
          $ref: "#/components/schemas/ChildBotConfigOverride"
    ChildBotConfigOverride:
      type: object
      properties:
        openingRangeMinutes:
          type: integer
        riskPercent:
          type: number
        rewardRiskRatio:
          type: number
        requireVolumeConfirmation:
          type: boolean
        volumeMultiplier:
          type: number
        trailingStopEnabled:
          type: boolean
        trailingStopActivationR:
          type: number
        reEntryEnabled:
          type: boolean
        maxOrbWidthPercent:
          type: number
        minOrbWidthPercent:
          type: number
        breakoutWindowMinutes:
          type: integer
        useModerateRisk:
          type: boolean
    ChildBotConfig:
      type: object
      required: [openingRangeMinutes, riskPercent, rewardRiskRatio, requireVolumeConfirmation, volumeMultiplier, trailingStopEnabled, trailingStopActivationR, reEntryEnabled, maxOrbWidthPercent, minOrbWidthPercent, breakoutWindowMinutes, useModerateRisk]
      properties:
        openingRangeMinutes:
          type: integer
        riskPercent:
          type: number
        rewardRiskRatio:
          type: number
        requireVolumeConfirmation:
          type: boolean
        volumeMultiplier:
          type: number
        trailingStopEnabled:
          type: boolean
        trailingStopActivationR:
          type: number
        reEntryEnabled:
          type: boolean
        maxOrbWidthPercent:
          type: number
        minOrbWidthPercent:
          type: number
        breakoutWindowMinutes:
          type: integer
        useModerateRisk:
          type: boolean
    BotStats:
      type: object
      required: [totalTrades, wins, losses, totalPnl, avgRMultiple, winRate]
      properties:
        totalTrades:
          type: integer
        wins:
          type: integer
        losses:
          type: integer
        totalPnl:
          type: number
        avgRMultiple:
          type: number
        winRate:
          type: number
    ChildBot:
      type: object
      required: [id, symbol, generation, config, phase, session, startedAt, stats]
      properties:
        id:
          type: string
        dbId:
          type: integer
          nullable: true
        symbol:
          type: string
        generation:
          type: integer
        parentId:
          type: string
          nullable: true
        config:
          $ref: "#/components/schemas/ChildBotConfig"
        phase:
          type: string
          enum: [idle, waiting_open, building_range, watching, in_trade, closed]
        session:
          $ref: "#/components/schemas/SessionState"
        startedAt:
          type: string
        stoppedAt:
          type: string
          nullable: true
        lastUpdated:
          type: string
          nullable: true
        error:
          type: string
          nullable: true
        stats:
          $ref: "#/components/schemas/BotStats"
    ScanCandidate:
      type: object
      required: [symbol, prevClose, open, currentPrice, gapPercent, gapDirection, volume, avgVolume, relativeVolume, recommended]
      properties:
        symbol:
          type: string
        prevClose:
          type: number
        open:
          type: number
        currentPrice:
          type: number
        gapPercent:
          type: number
        gapDirection:
          type: string
          enum: [up, down]
        volume:
          type: number
        avgVolume:
          type: number
        relativeVolume:
          type: number
        recommended:
          type: boolean
    ScanResult:
      type: object
      required: [scannedAt, candidates, topN]
      properties:
        scannedAt:
          type: string
          nullable: true
        candidates:
          type: array
          items:
            $ref: "#/components/schemas/ScanCandidate"
        topN:
          type: array
          items:
            $ref: "#/components/schemas/ScanCandidate"
    ScanRequest:
      type: object
      properties:
        symbols:
          type: array
          items:
            type: string
        topN:
          type: integer
          default: 10
        minGapPercent:
          type: number
          default: 0.5
        minRelativeVolume:
          type: number
          default: 1.0
...
    SessionState:
      type: object
      required: [symbol, phase, date]
      properties:
        symbol:
          type: string
        phase:
          type: string
          enum: [idle, waiting_open, building_range, watching, in_trade, closed]
        date:
          type: string
        orbHigh:
          type: number
          nullable: true
        orbLow:
          type: number
          nullable: true
        orbWidth:
          type: number
          nullable: true
        currentPrice:
          type: number
          nullable: true
        volume:
          type: number
          nullable: true
        averageVolume:
          type: number
          nullable: true
        volumeRatio:
          type: number
          nullable: true
        openRangeStart:
          type: string
          nullable: true
        openRangeEnd:
          type: string
          nullable: true
        breakoutDirection:
          type: string
          enum: [long, short, none]
          nullable: true
        entryPrice:
          type: number
          nullable: true
        stopPrice:
          type: number
          nullable: true
        targetPrice:
          type: number
          nullable: true
        currentPnl:
          type: number
          nullable: true
        longTradeUsed:
          type: boolean
        shortTradeUsed:
          type: boolean
        qty:
          type: number
          nullable: true
...