# lib/api-zod/src/generated/types/botStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Opening Range Breakout (ORB) Trading Bot API
 * OpenAPI spec version: 0.1.0
 */

export interface BotStats {
  totalTrades: number;
  wins: number;
  losses: number;
  totalPnl: number;
  avgRMultiple: number;
  winRate: number;
}