# lib/api-zod/src/generated/types/childBot.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Opening Range Breakout (ORB) Trading Bot API
 * OpenAPI spec version: 0.1.0
 */
import type { BotStats } from "./botStats";
import type { ChildBotConfig } from "./childBotConfig";
import type { ChildBotPhase } from "./childBotPhase";
import type { SessionState } from "./sessionState";

export interface ChildBot {
  id: string;
  dbId?: number | null;
  symbol: string;
  generation: number;
  parentId?: string | null;
  config: ChildBotConfig;
  phase: ChildBotPhase;
  session: SessionState;
  startedAt: string;
  stoppedAt?: string | null;
  lastUpdated?: string | null;
  error?: string | null;
  stats: BotStats;
}