# lib/api-zod/src/generated/types/childBotConfigOverride.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Opening Range Breakout (ORB) Trading Bot API
 * OpenAPI spec version: 0.1.0
 */

export interface ChildBotConfigOverride {
  openingRangeMinutes?: number;
  riskPercent?: number;
  rewardRiskRatio?: number;
  requireVolumeConfirmation?: boolean;
  volumeMultiplier?: number;
  trailingStopEnabled?: boolean;
  trailingStopActivationR?: number;
  reEntryEnabled?: boolean;
  maxOrbWidthPercent?: number;
  minOrbWidthPercent?: number;
  breakoutWindowMinutes?: number;
  useModerateRisk?: boolean;
}