# lib/api-zod/src/generated/types/createBotBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Opening Range Breakout (ORB) Trading Bot API
 * OpenAPI spec version: 0.1.0
 */
import type { ChildBotConfigOverride } from "./childBotConfigOverride";

export interface CreateBotBody {
  symbol: string;
  config?: ChildBotConfigOverride;
}