# lib/api-zod/src/generated/types/scanCandidate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Opening Range Breakout (ORB) Trading Bot API
 * OpenAPI spec version: 0.1.0
 */
import type { ScanCandidateGapDirection } from "./scanCandidateGapDirection";

export interface ScanCandidate {
  symbol: string;
  prevClose: number;
  open: number;
  currentPrice: number;
  gapPercent: number;
  gapDirection: ScanCandidateGapDirection;
  volume: number;
  avgVolume: number;
  relativeVolume: number;
  recommended: boolean;
}