# lib/api-zod/src/generated/types/scanResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Opening Range Breakout (ORB) Trading Bot API
 * OpenAPI spec version: 0.1.0
 */
import type { ScanCandidate } from "./scanCandidate";

export interface ScanResult {
  scannedAt: string | null;
  candidates: ScanCandidate[];
  topN: ScanCandidate[];
}