# lib/api-zod/src/generated/types/sessionState.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Opening Range Breakout (ORB) Trading Bot API
 * OpenAPI spec version: 0.1.0
 */
import type { SessionStateBreakoutDirection } from "./sessionStateBreakoutDirection";
import type { SessionStatePhase } from "./sessionStatePhase";

export interface SessionState {
  symbol: string;
  phase: SessionStatePhase;
  date: string;
  orbHigh?: number | null;
  orbLow?: number | null;
  orbWidth?: number | null;
  currentPrice?: number | null;
  volume?: number | null;
  averageVolume?: number | null;
  volumeRatio?: number | null;
  openRangeStart?: string | null;
  openRangeEnd?: string | null;
  breakoutDirection?: SessionStateBreakoutDirection;
  entryPrice?: number | null;
  stopPrice?: number | null;
  targetPrice?: number | null;
  currentPnl?: number | null;
  longTradeUsed?: boolean;
  shortTradeUsed?: boolean;
  qty?: number | null;
}